Extract status CLI command modules
