+++ regression/cli-command-module-contract.py
@@ -41,17 +41,25 @@ def main() -> int:
     sys.path.insert(0, str(REPO_ROOT))
     from goal_harness.cli_commands import (
         handle_demo_command,
+        handle_check_command,
         handle_doctor_command,
         handle_new_project_prompt_command,
+        handle_review_packet_command,
+        handle_status_command,
         register_doctor_command,
         register_starter_commands,
+        register_status_commands,
     )
 
     assert callable(handle_demo_command)
+    assert callable(handle_check_command)
     assert callable(register_doctor_command)
     assert callable(register_starter_commands)
+    assert callable(register_status_commands)
     assert callable(handle_doctor_command)
     assert callable(handle_new_project_prompt_command)
+    assert callable(handle_review_packet_command)
+    assert callable(handle_status_command)
 
     doctor = run_cli("doctor")
     assert doctor.get("ok") is True, doctor
@@ -84,6 +92,18 @@ def main() -> int:
     assert demo.get("ok") is True, demo
     assert demo.get("goal_id") == "command-module-demo", demo
     assert "quota" in demo, demo
+
+    check = run_cli("check", "--scan-root", "README.md", "--limit", "1")
+    assert check.get("ok") is True, check
+    assert "summary" in check, check
+
+    status = run_cli("status", "--limit", "1")
+    assert status.get("ok") is True, status
+    assert "attention_queue" in status, status
+
+    review = run_cli("review-packet", "--goal-id", "goal-harness-meta", "--limit", "1")
+    assert review.get("ok") is True, review
+    assert review.get("goal_id") == "goal-harness-meta", review
 
     print("cli-command-module-contract-regression ok")
     return 0
